requirements.txt
fastapi
uvicorn
python-dotenv
supabase
chromadb
google-generativeai
python-multipart
langchain-text-splitters
httpx
